backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0
fastmcp==0.1.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.3.7
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pymongo==4.6.0
motor==3.3.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pypdf2==3.0.1
tabula-py==2.8.2
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
scholarly==1.7.11
crossref-commons==0.0.7
citeproc-py==0.5.1
csl==0.1.0
firebase-admin==6.2.0
websockets==12.0
redis==5.0.1
celery==5.3.4 